+++ web/routes/routes.go
@@ -1,10 +1,13 @@
 package routes
 
 import (
-	"bitbucket.org/insamo/mvc/web/bootstrap"
+	"github.com/insamo/mvc/web/bootstrap"
+	"github.com/kataras/iris"
 )
 
 // Configure registers the necessary routes to the app.
 func Configure(b *bootstrap.Bootstrapper) {
-
+	b.Get("/ping", func(context iris.Context) {
+		return
+	})
 }